python/gh-issues-to-clickup/requirements.txt
PyGithub
requests
scramjet-framework-py
